# swagger/SocialnetApiDocs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Socialnet API",
    "description": "Code Lounge",
    "version": "1.6"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/account/me": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Получение аккаунта",
        "operationId": "getAccount",
        "responses": {
          "200": {
            "description": "Аккаунт успешно получен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "account-controller"
        ],
        "summary": "Обновление аккаунта",
        "operationId": "updateAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Аккаунт успешно обновлен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "account-controller"
        ],
        "summary": "Удаление аккаунта",
        "operationId": "deleteAccount",
        "responses": {
          "200": {
            "description": "Аккаунт успешно удален",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/account/": {
      "post": {
        "tags": [
          "account-controller"
        ],
        "summary": "Создание аккаунта",
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Аккаунт успешно создан",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/notifications/count": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "summary": "Получение количества уведомлений",
        "operationId": "getNotificationsCount",
        "responses": {
          "200": {
            "description": "Количество уведомлений успешно получено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DataTimestampDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/friends/count": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Получение количества входящих запросов в друзья",
        "operationId": "getIncomingFriendRequests",
        "responses": {
          "200": {
            "description": "Успешное получение количества запросов",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/account/{id}": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Получение аккаунта по ID",
        "operationId": "getAccountById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID аккаунта",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Аккаунт успешно получен по ID",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/account/search": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Поиск аккаунта",
        "operationId": "findAccount",
        "parameters": [
          {
            "name": "accountSearchDto",
            "in": "query",
            "description": "Критерии поиска",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AccountSearchDto"
            }
          },
          {
            "name": "pageDto",
            "in": "query",
            "description": "Информация о странице",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Аккаунт(ы) успешно найден(ы)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/account/search/statusCode": {
      "get": {
        "tags": [
          "account-controller"
        ],
        "summary": "Поиск аккаунтов по статус-коду отношений",
        "operationId": "findAccountByStatusCode",
        "parameters": [
          {
            "name": "accountSearchDto",
            "in": "query",
            "description": "Критерии поиска",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AccountSearchDto"
            }
          },
          {
            "name": "pageDto",
            "in": "query",
            "description": "Информация о странице",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Аккаунт(ы) успешно найден(ы)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор аккаунта",
            "format": "uuid"
          },
          "isDeleted": {
            "type": "boolean",
            "description": "Статус удаления аккаунта"
          },
          "firstName": {
            "type": "string",
            "description": "Имя пользователя"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия пользователя"
          },
          "email": {
            "type": "string",
            "description": "Электронная почта пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "writeOnly": true
          },
          "phone": {
            "type": "string",
            "description": "Телефон пользователя"
          },
          "about": {
            "type": "string",
            "description": "Описание пользователя"
          },
          "city": {
            "type": "string",
            "description": "Город пользователя"
          },
          "country": {
            "type": "string",
            "description": "Страна пользователя"
          },
          "statusCode": {
            "type": "string",
            "description": "Код статуса",
            "enum": [
              "FRIEND",
              "REQUEST_TO",
              "REQUEST_FROM",
              "BLOCKED",
              "DECLINED",
              "SUBSCRIBED",
              "NONE",
              "WATCHING",
              "REJECTING",
              "RECOMMENDATION"
            ]
          },
          "regDate": {
            "pattern": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
            "type": "string",
            "description": "Дата регистрации аккаунта",
            "format": "date-time"
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения пользователя",
            "format": "date"
          },
          "messagePermission": {
            "type": "boolean",
            "description": "Разрешение на отправку сообщений"
          },
          "lastOnlineTime": {
            "pattern": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
            "type": "string",
            "description": "Последнее время онлайн",
            "format": "date-time"
          },
          "isOnline": {
            "type": "boolean",
            "description": "Статус онлайн"
          },
          "isBlocked": {
            "type": "boolean",
            "description": "Статус блокировки аккаунта"
          },
          "photo": {
            "type": "string",
            "description": "Фотография профиля"
          },
          "profileCover": {
            "type": "string",
            "description": "Обложка профиля"
          },
          "createdOn": {
            "pattern": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
            "type": "string",
            "description": "Дата создания аккаунта",
            "format": "date-time"
          },
          "updatedOn": {
            "pattern": "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
            "type": "string",
            "description": "Дата обновления аккаунта",
            "format": "date-time"
          }
        },
        "description": "Данные аккаунта"
      },
...
      "PageableDto": {
        "type": "object",
        "properties": {
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageNumber": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "description": "Смещение",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean",
            "description": "Без разбивки на страницы"
          },
          "paged": {
            "type": "boolean",
            "description": "Разбито на страницы"
          }
        },
        "description": "Данные для постраничного вывода"
      },
      "SortDto": {
        "type": "object",
        "properties": {
          "unsorted": {
            "type": "boolean",
            "description": "Не отсортировано"
          },
          "sorted": {
            "type": "boolean",
            "description": "Отсортировано"
          },
          "empty": {
            "type": "boolean",
            "description": "Пусто"
          }
        },
        "description": "Данные сортировки"
      },
...
      "DataTimestampDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Временна́я метка",
            "format": "date-time"
          },
          "data": {
            "type": "object",
            "description": "Данные"
          }
        },
        "description": "Данные для уведомления с timestamp"
      },
...
      "PageDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Страница (номер)",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "description": "Размер",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "description": "Сортировка",
            "items": {
              "type": "string",
              "description": "Сортировка"
            }
          }
        },
        "description": "Данные страницы"
      },
...
      "CountDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "description": "Количество",
            "format": "int32"
          }
        },
        "description": "Счетчик"
      },
...
      "AccountSearchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID аккаунта",
            "format": "uuid"
          },
          "ids": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "description": "Список ID аккаунтов",
            "example": [
              "00000000-0000-0000-0000-000000000000",
              "11111111-1111-1111-1111-111111111111"
            ],
            "items": {
              "type": "string"
            }
          },
          "blockedByIds": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "description": "ID заблокировавших аккаунтов",
            "example": [
              "00000000-0000-0000-0000-000000000000",
              "11111111-1111-1111-1111-111111111111"
            ],
            "items": {
              "type": "string"
            }
          },
          "author": {
            "type": "string",
            "description": "Автор аккаунта"
          },
          "firstName": {
            "type": "string",
            "description": "Имя аккаунта"
          },
          "lastName": {
            "type": "string",
            "description": "Фамилия аккаунта"
          },
          "maxBirthDate": {
            "type": "string",
            "description": "Максимальная дата рождения",
            "format": "date"
          },
          "minBirthDate": {
            "type": "string",
            "description": "Минимальная дата рождения",
            "format": "date"
          },
          "city": {
            "type": "string",
            "description": "Город"
          },
          "country": {
            "type": "string",
            "description": "Страна"
          },
          "statusCode": {
            "type": "string",
            "description": "Код статуса",
            "enum": [
              "FRIEND",
              "REQUEST_TO",
              "REQUEST_FROM",
              "BLOCKED",
              "DECLINED",
              "SUBSCRIBED",
              "NONE",
              "WATCHING",
              "REJECTING",
              "RECOMMENDATION"
            ]
          },
          "ageTo": {
            "type": "integer",
            "description": "Максимальный возраст",
            "format": "int32"
          },
          "ageFrom": {
            "type": "integer",
            "description": "Минимальный возраст",
            "format": "int32"
          },
          "deleted": {
            "type": "boolean"
          },
          "blocked": {
            "type": "boolean"
          }
        },
        "description": "Форма поиска аккаунта"
      },
      "PageResponseDtoAccountDto": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "description": "Общее количество элементов",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "description": "Общее количество страниц",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "description": "Номер страницы",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "description": "Размер страницы",
            "format": "int32"
          },
          "first": {
            "type": "boolean",
            "description": "Первая страница"
          },
          "last": {
            "type": "boolean",
            "description": "Последняя страница"
          },
          "numberOfElements": {
            "type": "integer",
            "description": "Количество элементов на странице",
            "format": "int32"
          },
          "empty": {
            "type": "boolean",
            "description": "Пустая страница"
          },
          "sort": {
            "$ref": "#/components/schemas/SortDto"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableDto"
          },
          "content": {
            "minItems": 0,
            "uniqueItems": true,
            "type": "array",
            "description": "Содержимое страницы",
            "items": {
              "$ref": "#/components/schemas/AccountDto"
            }
          }
        },
        "description": "Ответ с данными страницы"
      }
    }
  }
}